pyscripts/softswitch/requirements.txt
# typing==3.6.6
django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
# djangorestframework-bulk==0.2.1
# South==1.0.2
coreapi==2.3.3
# simple-rest-client==0.5.4
# djmicroservice>=1.0
# pyapollo>=0.0.1
pyjwt-rsa==0.3.0
# markdown==3.0.1
django-cors-middleware==1.3.1
pyyaml==3.13
django-cors-headers==3.1.1
mysqlclient==1.4.4

# uWSGI==2.0.18